Remove prices from everything
Prices will change from time to time, and this site is meant to be extremely low-maintencance. Prices should not display.
Also removed product-grid layout, because the only point of that was to create a separation between grid layouts where one would not show prices (category-grid) and one would show prices (product-grid). Prices are no longer a thing, so now there is only one, called product-grid